FastAPI/requirements.txt
uvicorn
fastapi
mysqlclient
mysql-connector-python
python-dotenv
